\name{datagen.pleiomap}
\alias{datagen.pleiomap}
\title{Simulation Function for the STARSEQ method}
\description{
  Simulating Data for Evaluating STARSEQ package
}
\usage{
datagen.pleiomap(pars)
}
\arguments{
  \item{pars}{
    A list of simulation parameters
  }
}
\value{
  \item{dat.ped}{ A list of simulated ped file}
  \item{y2.vec}{Simulated secondary traits}
  \item{sumstat}{Summary statistics calculated for simulated dataset}
}
\author{
  Dajiang Liu
}
