\name{kbac.weight}
\alias{kbac.weight}

\title{Weighting Function for KBAC
}
\description{
  Weighting function for KBAC, does not take user input
}
\usage{
kbac.weight(dat)
}
\arguments{
  \item{dat}{ Matrix format for genotype and phenotypes.
  }
}
\value{
  \item{x}{recoded multi-site genotypes}
}
\author{
  Dajiang Liu
}
