\name{lse.wss.misshsq}
\alias{lse.wss.misshsq}
\title{ Least Square Estimates for Genetic Average Effect and Locus
  Heritability in the Weighted Sum Statistics
}
\description{
  Least Square Estimates for Genetic Average Effect and Locus
  Heritability in the Weighted Sum Statistics
}
\usage{
lse.wss.misshsq(marker, y, weight)
}

\arguments{
  \item{marker}{
    Genotype Matrix
  }
  \item{y}{
    Quantitative trait
  }
  \item{weight}{
    Weight assigned to each variant site;
  }
}

\value{
  \item{beta1.est}{Estimates of locus genetic effects;}
  \item{hsq.est}{Estimates of locus genetic variance}
}

\author{
  Dajiang Liu
}
