\name{mat2ped}
\alias{mat2ped}

\title{
  Transform Mat format to PED format
}
\description{
  Transform Mat format data to ped format. Internal function; Not
  designed for user input;
}
\usage{
mat2ped(dat.mat)
}

\arguments{
  \item{dat.mat}{
    Data in mat format
  }
}
\value{
  Ped format output
}
\author{
  Dajiang Liu 
}
