\name{recode.uniqtl}
\alias{recode.uniqtl}

\title{Recode Multi-site Genotypes for KBAC Test

}
\description{
Record Multi-site Genotypes for KBAC Test
}
\usage{
recode.uniqtl(dat)
}

\arguments{
  \item{dat}{
    Input data in matrix format;
  }
}
\value{
  Return recoded genotypes in mat format;
}

\author{
  Dajiang Liu
}
