\name{set.compare}
\alias{set.compare}
\title{Comparing 2 Ordered Set of Values
}
\description{
Compare if two ordered sets are equivalent;
}
\usage{
set.compare(set1, set2)
}

\arguments{
  \item{set1}{
    The first set;
  }
  \item{set2}{
    The second set;
  }
}

\value{
 1 if the two sets are equal, 0 if not;
}

\author{
  Dajiang Liu
}
