\name{set.intersect}
\alias{set.intersect}
\title{
  Calculate Intersection of Two Sets
}
\description{
  Calculate Intersection of Two Sets
}
\usage{
set.intersect(set1, set2)
}
\arguments{
  \item{set1}{
    A vector of set 1
  }
  \item{set2}{
    A vector of set 2
  }
}
\value{
  An intersection of Two Sets.
}
\author{
  Dajiang Liu
}
