\name{STAR-package}
\alias{STAR-package}
\alias{STAR}
\docType{package}
\title{
  Secondary Trait Association analysis for Rare variants via SEQuence data
}
\description{
  STARSEQ is an R-package for detecting associations with rare variants using selected samples in sequence-based association studies. It corrects for the bias in the secondary trait distribution induced by selective sampling on the primary trait. The corrected secondary trait can be analyzed by standard rare variant tests. In the STARSEQ package, several popular rare variant tests were implemented, which include 1.) combined multivariate and collapsing 2.) weighted sum statistics, 3.) kernel based adaptive cluster, 4.) variable threshold test 5.) sequence kernel association test. 
}
\details{
  \tabular{ll}{
    Package: \tab STAR\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2012-05-15\cr
    License: \tab GPL-3 \cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Dajiang Liu
  Maintainer: Dajiang Liu <dajiang.liu@gmail.com>
}
\references{
  1. Liu, D. J. and S. M. Leal (2011). "A flexible likelihood framework for
  detecting associations with secondary phenotypes in genetic studies
  using selected samples: application to sequence data." Eur J Hum
  Genet.
  
  2. Liu, D. J. and S. M. Leal A Unified Method for Detecting Secondary Trait
  Associations with Rare Variants: Application to Sequence Data PLoS
  Genetics, under review
}

