\name{std.score.proscore.stat.mat}
\alias{std.score.proscore.stat.mat}
\title{ Single Variant Score Statistics for a Genotype Matrix
}
\description{
  Single Variant Score Statistics for a Genotype Matrix
}
\usage{
std.score.proscore.stat.mat(x.mat, y, yub, ylb, par.nuis)
}
\arguments{
  \item{x.mat}{
    Genotype matrix
}
  \item{y}{
    Quantitative or binary trait
}
  \item{yub}{
    Upper trait threshold
}
  \item{ylb}{
    Lower trait threshold
}
  \item{par.nuis}{
    Nuissance parameters
  }
}
\value{
The function returns an array of statistics 
}
\author{
  Dajiang Liu
}
