\name{std.score.stat.new}
\alias{std.score.stat.new}
\title{
  Standard Score Statistics from Prospective Likelihood
}
\description{
  Prospective Likelihood Score Statistics
}
\usage{
std.score.stat.new(x, y, yub, ylb, par.nuis)
}
\arguments{
  \item{x}{
    Predictor variable;
  }
  \item{y}{
    Response variable.
}
  \item{yub}{
    Upper trait threshold. 
}
  \item{ylb}{
    Lower trait threshold;
}
  \item{par.nuis}{
    Other nuissance parameters;
  }
}
\value{
  Value of statistics from the prospective likelihhod
}
\author{
  Dajiang Liu
}